+++ examples/react-mobx-ts/src/views/CustomLayout/ModeratorViewModel.ts
@@ -1,0 +1,30 @@
+import { action, computed, observable, reaction, when } from "mobx";
+import Managers from "../../stores/Managers";
+import { BJNWebClientSDK, StreamQuality, StreamPriority, videoStream, Participant } from '@bluejeans/web-client-sdk';
+import CustomLayoutManager from "../../stores/CustomLayoutManager"
+import CustomLayoutViewModel from "./CustomLayoutViewModel";
+
+export default class ModeratorViewModel {
+
+    private webrtcSDK: BJNWebClientSDK;
+    private customLayoutManager: CustomLayoutManager
+    private customLayoutViewModel: CustomLayoutViewModel
+
+    constructor(managers: Managers) {
+        this.webrtcSDK = managers.webrtcSDK;
+        this.customLayoutManager = managers.customLayoutManager
+        this.customLayoutViewModel = new CustomLayoutViewModel(managers)
+    }
+
+    renderParticipants(participantGuid, element) {
+        this.webrtcSDK.meetingService.videoStreamService?.attachParticipantToView(participantGuid, <HTMLDivElement>document.getElementById(`${element}`))
+    }
+
+    @computed get videoStreamsMap(): Map<string, videoStream> {
+        return this.customLayoutManager.videoStreamsMap;
+    }
+
+    initials(participant): string {
+        return this.customLayoutManager.getInitials(participant)
+    }
+}